Added a "trigger type" for sensors.
Added the ability to define what events cause the sensor to be "triggered". Added the "trigger.playStop" type which is only triggered on by play events and off by stop events. This deprecates the 'ignorePauseResume' variable. Added the "trigger.pauseResume" which is only triggered on by pause events and off by resume event. Added the "trigger.scrobble" event which is triggered on when the media reaches 90% play through. Added the "trigger.normal" type which is the default type for all sensors and is triggered on for play and resume events and trigger off by pause and stop events.